restructure the project
- refactor and remove multiple usage of IO operations in the processs
- chage the structure of the project
