+++ cmd/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"cdc-file-processor/domain"
+)
+
+func main() {
+	domain.Cdc_run()
+}
