+++ files/input.txt
@@ -1,3 +1,5 @@
 sdgbsfdbs
 sdgds
 changed
+chanhe
+SDgSD
